BumdesKM

+++ AndroidAPI/update_foto.php
@@ -8,15 +8,17 @@ if ($_SERVER["REQUEST_METHOD"] == 'POST') {
 
     $getMe = mysqli_fetch_assoc(mysqli_query($conn, "SELECT * FROM users WHERE id_user = '$id_user' "));
     $oldPhoto = $getMe['foto_profil'];
-    $file = "../img/user/".$oldPhoto;
+    $file = "../images/user/".$oldPhoto;
     
     $filename = uniqid().".jpeg";
     
     mysqli_query($conn, "UPDATE users SET foto_profil = '$filename' WHERE id_user = '$id_user'");
     
     if (mysqli_affected_rows($conn) > 0) {
-        file_put_contents("../img/user/".$filename, base64_decode($encoded_file));
-        unlink($file);
+        file_put_contents("../images/user/".$filename, base64_decode($encoded_file));
+        if ($oldPhoto != 'default_user.png') {
+          unlink($file);
+        }
         $response = array('pesan' => 'BERHASIL');
     } else {
         $response = array('pesan' => 'GAGAL');
